+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/Common.kt
@@ -55,7 +55,9 @@ val botCommandLimit = botCommandLengthLimit
 val botCommandDescriptionLimit = 3 .. 256
 val botCommandsLimit = 0 .. 100
 
-val quizPollExplanationLimit = 0 .. 200
+val explanationLimit = 0 .. 200
+@Deprecated("Will be removed in near updates", ReplaceWith("explanationLimit"))
+val quizPollExplanationLimit = explanationLimit
 
 val openPeriodPollSecondsLimit = 5 .. 600